enh(a11y): User management settings dialog
